refactor compose auth
