+++ plugins/ComposeAuth/src/appleMain/kotlin/io/github/jan/supabase/compose/auth/composable/GoogleAuth.kt
@@ -8,10 +8,10 @@ import io.github.jan.supabase.gotrue.SignOutScope
  * Composable for Google login with default behavior
  */
 @Composable
-actual fun ComposeAuth.rememberLoginWithGoogle(onResult: (NativeSignInResult) -> Unit, fallback: suspend () -> Unit): NativeSignInState = defaultLoginBehavior(fallback)
+actual fun ComposeAuth.rememberSignInWithGoogle(onResult: (NativeSignInResult) -> Unit, fallback: suspend () -> Unit): NativeSignInState = defaultLoginBehavior(fallback)
 
 /**
  * Composable for SignOut with default behavior
  */
 @Composable
-actual fun ComposeAuth.rememberSignOut(signOutScope: SignOutScope): NativeSignInState = defaultSignOutBehavior(signOutScope)
+actual fun ComposeAuth.rememberSignOutWithGoogle(signOutScope: SignOutScope): NativeSignInState = defaultSignOutBehavior(signOutScope)
