+++ plugins/ComposeAuth/src/commonMain/kotlin/io/github/jan/supabase/compose/auth/composable/NativeGoogleAuth.kt
@@ -1,11 +1,8 @@
 package io.github.jan.supabase.compose.auth.composable
 
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.runtime.remember
 import io.github.jan.supabase.compose.auth.ComposeAuth
 import io.github.jan.supabase.compose.auth.fallbackLogin
-import io.github.jan.supabase.compose.auth.signOut
 import io.github.jan.supabase.gotrue.SignOutScope
 import io.github.jan.supabase.gotrue.providers.Google
 
@@ -14,42 +11,20 @@ import io.github.jan.supabase.gotrue.providers.Google
  * @return [NativeSignInState]
  */
 @Composable
-expect fun ComposeAuth.rememberLoginWithGoogle(onResult: (NativeSignInResult) -> Unit = {}, fallback: suspend () -> Unit = { fallbackLogin(Google) }): NativeSignInState
+expect fun ComposeAuth.rememberSignInWithGoogle(onResult: (NativeSignInResult) -> Unit = {}, fallback: suspend () -> Unit = { fallbackLogin(Google) }): NativeSignInState
 
 /**
- * Composable of default behavior if Native Login is not supported on the platform
+ * Composable function what implements Native Google Login flow with default fallback
+ * @return [NativeSignInState]
  */
 @Composable
-fun defaultLoginBehavior(fallback: suspend () -> Unit): NativeSignInState {
-    val state = remember { NativeSignInState() }
-    LaunchedEffect(key1 = state.started) {
-        if (state.started) {
-            fallback.invoke()
-            state.reset()
-        }
-    }
-    return state
-}
+@Deprecated("Use rememberSignInWithGoogle instead", ReplaceWith("rememberSignInWithGoogle(onResult, fallback)"), DeprecationLevel.WARNING)
+fun ComposeAuth.rememberLoginWithGoogle(onResult: (NativeSignInResult) -> Unit = {}, fallback: suspend () -> Unit = { fallbackLogin(Google) }): NativeSignInState = rememberSignInWithGoogle(onResult, fallback)
 
 /**
  * Composable for sign out flow
  * @return [NativeSignInState]
  */
 @Composable
-expect fun ComposeAuth.rememberSignOut(signOutScope: SignOutScope = SignOutScope.LOCAL): NativeSignInState
+expect fun ComposeAuth.rememberSignOutWithGoogle(signOutScope: SignOutScope = SignOutScope.LOCAL): NativeSignInState
 
-/**
- * Composable used to Sign Out from GoTrue
- */
-@Composable
-fun ComposeAuth.defaultSignOutBehavior(signOutScope: SignOutScope, nativeSignOut: suspend () -> Unit = {}): NativeSignInState {
-    val state = remember { NativeSignInState() }
-    LaunchedEffect(key1 = state.started) {
-        if (state.started) {
-            nativeSignOut.invoke()
-            signOut(signOutScope)
-            state.reset()
-        }
-    }
-    return state
-}
