+++ plugins/ComposeAuth/src/commonMain/kotlin/io/github/jan/supabase/compose/auth/composable/NativeSignInResult.kt
@@ -14,7 +14,7 @@ sealed interface NativeSignInResult {
 
     /**
      *
-     * User canceled or clicked away
+     * User cancelled or clicked away
      *
      */
     data object ClosedByUser : NativeSignInResult
