+++ plugins/ComposeAuth/src/jvmMain/kotlin/io/github/jan/supabase/compose/auth/composable/AppleAuth.kt
@@ -7,4 +7,4 @@ import io.github.jan.supabase.compose.auth.ComposeAuth
  * Composable for Apple login with default behavior
  */
 @Composable
-actual fun ComposeAuth.rememberLoginWithApple(onResult: (NativeSignInResult) -> Unit, fallback: suspend () -> Unit): NativeSignInState = defaultLoginBehavior(fallback)
+actual fun ComposeAuth.rememberSignInWithApple(onResult: (NativeSignInResult) -> Unit, fallback: suspend () -> Unit): NativeSignInState = defaultLoginBehavior(fallback)
